Moved FileList and FileListItem components into their own directory with their own styles file
