+++ packages/veritone-react-common/src/components/FilePicker/FileList/FileListItem.js
