+++ packages/veritone-react-common/src/components/FilePicker/styles.scss
@@ -1,13 +1,6 @@
 @import 'src/styles/modules/variables';
 @import 'src/styles/modules/muiTypography';
 
-@mixin file-picker-round-image($size: 40px, $margin: 5px) {
-    width: $size;
-    height: $size;
-    margin: $margin;
-    border-radius: 50%;
-}
-
 .filePicker {
     display: -webkit-flex;
     display: flex;
@@ -55,48 +48,6 @@
 .fileUploaderSubtext {
     @include mui-text('subhead');
     margin-bottom: $file-picker-margin;
-}
-
-.fileList {
-    flex: 0 0 50%;
-}
-
-.fileListItem {
-    display: -webkit-flex;
-    display: flex;
-    -webkit-flex-direction: row;
-    flex-direction: row;
-    height: 50px;
-}
-
-.fileListItemText {
-    overflow: hidden;
-    text-overflow: ellipsis;
-    display:block;
-    padding: 4px;
-    span {
-        width: 100%;
-        white-space:nowrap;
-        display: inline-block;
-    }
-}
-
-.fileListItemNameText {
-    @include mui-text('body1');
-}
-
-.fileListItemFileSizeText {
-    @include mui-text('body1');
-    color: $grey-5;
-}
-
-.fileListItemDeleteIcon {
-    margin-left: auto;
-}
-
-.fileListItemImage {
-    @include file-picker-round-image(40px, 4px);
-    border-radius: 50%;
 }
 
 .urlUploader {
